script changes

+++ scripts/encrypt.sh
@@ -5,6 +5,13 @@ TOKEN_DIR="${HOME}/.tokens"
 diff "${TOKEN_DIR}/${TOKEN}" "${TOKEN}"
 NEW_TOKEN=$?
 
+if [[ $CI != true ]]; then
+    # Load env vars
+    set -a
+    . ./.env
+    set +a
+fi
+
 # If Robinhood API grants us a new token:
 if [[ ${NEW_TOKEN} == 1 ]]; then 
     # Encrypt token
@@ -12,5 +19,7 @@ if [[ ${NEW_TOKEN} == 1 ]]; then
 fi
 
 # Remove leftover tokens
-rm -rf "${TOKEN_DIR}"
+if [[ $CI == true ]]; then
+    rm -rf "${TOKEN_DIR}"
+fi
 rm "${TOKEN}"
